Prompt Engineering Playground/requirements.txt
gradio
google-generativeai
python-dotenv